+++ packages/client/signers/src/api/schemas/completeOnboarding.ts
@@ -1,17 +1,5 @@
 import { z } from "zod";
 
-export const startOnboardingRequestSchema = z.object({
-    authId: z.string(),
-    deviceId: z.string(),
-    encryptionContext: z.object({
-        publicKey: z.string(),
-    }),
-});
-export type StartOnboardingRequest = z.infer<typeof startOnboardingRequestSchema>;
-
-export const startOnboardingResponseSchema = z.object({});
-export type StartOnboardingResponse = z.infer<typeof startOnboardingResponseSchema>;
-
 export const completeOnboardingRequestSchema = z.object({
     publicKey: z.string(),
     onboardingAuthentication: z.object({
