+++ packages/client/signers/src/api/schemas/getAuthShard.ts
