+++ packages/client/signers/src/api/schemas/index.ts
@@ -1,3 +1,5 @@
-export * from "./auth";
-export * from "./attestation";
-export * from "./onboarding";
+export * from "./completeOnboarding";
+export * from "./getAttestation";
+export * from "./getAuthShard";
+export * from "./getPublicKey";
+export * from "./startOnboarding";
